feat: modularize history page and fix the themed style
